back/requirements.txt
sqlalchemy
fastapi[standard]
python-dotenv
asyncpg
pyjwt
passlib[bcrypt]
pydantic[email]